requirements.txt
streamlit
pdfplumber
numpy
requests
faiss-cpu
python-dotenv
openai